requirements.txt
pandas
requests
pymmwr
click
urllib3
selenium
webdriver-manager
pyyaml
PyGithub
pytz